+++ relay/nakama/auth/custom_test.go
@@ -1,4 +1,4 @@
-package main
+package auth
 
 import (
 	"context"
